+++ libs/feat-utils/.mocharc.js
@@ -1,0 +1,6 @@
+require('@swc-node/register');
+
+module.exports = {
+    timeout: 60000,
+    require: ['should'],
+};