+++ libs/jsonv/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @galaxar/jsonv
 
+## 1.0.2
+
+### Patch Changes
+
+-   Fix bugs.
+-   Updated dependencies
+    -   @galaxar/types@1.0.2
+    -   @galaxar/utils@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes