+++ libs/jsonx/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @galaxar/jsonx
 
+## 1.0.2
+
+### Patch Changes
+
+-   Fix bugs.
+-   Updated dependencies
+    -   @galaxar/jsonv@1.0.2
+    -   @galaxar/types@1.0.2
+    -   @galaxar/utils@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes