+++ libs/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @galaxar/types
 
+## 1.0.2
+
+### Patch Changes
+
+-   Fix bugs.
+-   Updated dependencies
+    -   @galaxar/utils@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes