+++ libs/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @galaxar/utils
 
+## 1.0.2
+
+### Patch Changes
+
+-   Fix bugs.
+
 ## 1.0.1
 
 ### Patch Changes